+++ .env.example
@@ -12,6 +12,4 @@ BUCKET=
 SENTRY=
 ENVIRONMENT=
 
-ORG=
-
 SENTRY_AUTH_TOKEN=
