+++ discoverAPI.go
@@ -26,14 +26,7 @@ type EventMetadata struct {
 // Events from last 24HrPeriod events for selected Projects
 // Returns event metadata (e.g. Id, Project) but not the entire Event itself, which gets queried separately.
 func (d DiscoverAPI) latestEventMetadata(org string, n int) []EventMetadata {
-	// org := os.Getenv("ORG")
-	fmt.Println("> org", org)
-
-	// query := "&query=platform.name%3Ajavascript+OR+platform.name%3Apython"
 	query := "platform.name%3Ajavascript+OR+platform.name%3Apython"
-
-	// TEST
-	// endpoint := fmt.Sprintf("https://sentry.io/api/0/organizations/%v/eventsv2/?statsPeriod=24h&field=event.type&field=project&field=platform.name&sort=-event.type&per_page=25%v", org, query)
 
 	// with 0 project names specified
 	endpoint := fmt.Sprintf("https://sentry.io/api/0/organizations/%v/eventsv2/?statsPeriod=24h&field=event.type&field=project&field=platform&per_page=%v&query=%v", org, strconv.Itoa(n), query)
